+++ CHANGELOG.md
@@ -9,6 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Nothing yet.
 
+## [1.4.0] - 2023-08-13
+
+### Added
+
+- Implemented a RandomStringGenerator.
+- Added an AggregateId argument in Identity aggregate constructors.
+
 ## [1.3.0] - 2023-08-13
 
 ### Added
@@ -55,7 +62,8 @@ Nothing yet.
 
 - Implemented StringExtensions.
 
-[unreleased]: https://github.com/Logitar/Logitar.NET/compare/v1.3.0...HEAD
+[unreleased]: https://github.com/Logitar/Logitar.NET/compare/v1.4.0...HEAD
+[1.4.0]: https://github.com/Logitar/Logitar.NET/releases/tag/v1.3.0...v1.4.0
 [1.3.0]: https://github.com/Logitar/Logitar.NET/releases/tag/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/Logitar/Logitar.NET/releases/tag/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/Logitar/Logitar.NET/releases/tag/v1.0.0...v1.1.0